# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Generic events ingest service for durable event envelopes.",
        "title": "sigint API",
        "contact": {
            "name": "Applied Symbolics"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8920",
    "basePath": "/",
    "paths": {
...
        "/internal/v1/events/replay": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "replays stored events in ingest order using an exclusive cursor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "replay stored events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exclusive replay cursor",
                        "name": "after_cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum events to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Producer service filter",
                        "name": "producer_service",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Event name filter",
                        "name": "event_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject type filter",
                        "name": "subject_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject id filter",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Aggregate type filter",
                        "name": "aggregate_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Aggregate id filter",
                        "name": "aggregate_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Correlation id filter",
                        "name": "correlation_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/events.ReplayPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "validates canonical hashes, writes the raw archive, and records ingest state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "ingest one event",
                "parameters": [
                    {
                        "description": "Event envelope",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/events.Envelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ingestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/events/{event_id}": {
            "get": {
                "description": "returns the ingest ledger record and raw envelope for one event id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "get one event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.recordResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/events:batch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "validates and ingests 1 to 1000 event envelopes, returning one result per event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "ingest a batch of events",
                "parameters": [
                    {
                        "description": "Event batch",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/events.BatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.batchResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.batchResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ingestResponse"
                    }
                }
            }
        },
        "controller.ingestResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "received_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storage_uri": {
                    "type": "string"
                },
                "stored_at": {
                    "type": "string"
                }
            }
        },
        "controller.recordResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_sha256": {
                    "type": "string"
                },
                "event_version": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "partition_id": {
                    "type": "string"
                },
                "payload_sha256": {
                    "type": "string"
                },
                "producer_service": {
                    "type": "string"
                },
                "raw_envelope": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "received_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storage_uri": {
                    "type": "string"
                },
                "stored_at": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                }
            }
        },
        "events.Actor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "events.BatchRequest": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/events.Envelope"
                    }
                }
            }
        },
        "events.Envelope": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/events.Actor"
                },
                "aggregate_id": {
                    "type": "string"
                },
                "aggregate_type": {
                    "type": "string"
                },
                "causation_id": {
                    "type": "string"
                },
                "correlation_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_sha256": {
                    "type": "string"
                },
                "event_version": {
                    "type": "string"
                },
                "observed_at": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "partition_id": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "payload_sha256": {
                    "type": "string"
                },
                "producer_deployment": {
                    "type": "string"
                },
                "producer_instance": {
                    "type": "string"
                },
                "producer_service": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                }
            }
        },
        "events.ReplayEvent": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "aggregate_id": {
                    "type": "string"
                },
                "aggregate_type": {
                    "type": "string"
                },
                "causation_id": {
                    "type": "string"
                },
                "correlation_id": {
                    "type": "string"
                },
                "cursor": {
                    "type": "integer"
                },
                "event_id": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_sha256": {
                    "type": "string"
                },
                "event_version": {
                    "type": "string"
                },
                "observed_at": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "partition_id": {
                    "type": "string"
                },
                "payload_sha256": {
                    "type": "string"
                },
                "producer_deployment": {
                    "type": "string"
                },
                "producer_instance": {
                    "type": "string"
                },
                "producer_service": {
                    "type": "string"
                },
                "raw_envelope_json": {
                    "type": "string"
                },
                "received_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "storage_uri": {
                    "type": "string"
                },
                "stored_at": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                }
            }
        },
        "events.ReplayPage": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/events.ReplayEvent"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Optional bearer token for producer ingest and internal replay routes.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}